refactor(server): Create TestRuleSet per test class rather than share them #560
